check for instanceof and reorganize files to prevent circular dependencies
