--- packages/web-components/fast-components/src/design-system-consumer/index.ts
@@ -1,1 +1,0 @@
-export * from "./design-system-consumer";
