+++ packages/web-components/fast-components/src/design-system-provider/design-system-provider.styles.ts
@@ -1,5 +1,5 @@
 import { css } from "@microsoft/fast-element";
-import { display } from "../styles";
+import { display } from "../styles/display";
 
 export const DesignSystemProviderStyles = css`
     ${display("block")};
